+++ examples/messaging_recvonly_sample/src/messaging_recvonly_sample.cpp
@@ -140,7 +140,7 @@ int main(int argc, char* argv[]) {
 
   auto is_json = CLI::Validator(
       [](std::string input) -> std::string {
-        boost::json::error_code ec;
+        boost::system::error_code ec;
         boost::json::parse(input);
         if (ec) {
           return "Value " + input + " is not JSON Value";