+++ examples/sumomo/src/sumomo.cpp
@@ -1,7 +1,6 @@
 // Sora
 #include <sora/camera_device_capturer.h>
 #include <sora/sora_client_context.h>
-#include <sora/srtp_keying_material_exporter.h>
 
 #include <fstream>
 #include <optional>
@@ -27,6 +26,7 @@ struct SumomoConfig {
   std::string client_id;
   bool video = true;
   bool audio = true;
+  std::string video_device;
   std::string video_codec_type;
   std::string audio_codec_type;
   std::string resolution = "VGA";
@@ -54,12 +54,12 @@ struct SumomoConfig {
   bool show_me = false;
   bool fullscreen = false;
 
-  std::string srtp_key_log_file;
-
   bool insecure = false;
   std::string client_cert;
   std::string client_key;
   std::string ca_cert;
+
+  std::optional<webrtc::DegradationPreference> degradation_preference;
 
   struct Size {
     int width;
@@ -109,6 +109,7 @@ class Sumomo : public std::enable_shared_from_this<Sumomo>,
       cam_config.height = size.height;
       cam_config.fps = 30;
       cam_config.use_native = config_.hw_mjpeg_decoder;
+      cam_config.device_name = config_.video_device;
       auto video_source = sora::CreateCameraDeviceCapturer(cam_config);
       if (video_source == nullptr) {
         RTC_LOG(LS_ERROR) << "Failed to create video source.";
@@ -185,6 +186,7 @@ class Sumomo : public std::enable_shared_from_this<Sumomo>,
     if (!config_.ca_cert.empty()) {
       config.ca_cert = load_file(config_.ca_cert);
     }
+    config.degradation_preference = config_.degradation_preference;
     conn_ = sora::SoraSignaling::Create(config);
 
     boost::asio::executor_work_guard<boost::asio::io_context::executor_type>
@@ -227,36 +229,7 @@ class Sumomo : public std::enable_shared_from_this<Sumomo>,
     ioc_->stop();
   }
   void OnNotify(std::string text) override {
-    auto json = boost::json::parse(text);
-    if (json.at("event_type").as_string() == "connection.created") {
-      if (!config_.srtp_key_log_file.empty() && !key_exported_) {
-        auto km = sora::ExportKeyingMaterial(conn_->GetPeerConnection(),
-                                             conn_->GetVideoMid());
-        if (!km) {
-          RTC_LOG(LS_ERROR) << "Failed to ExportKeyingMaterial";
-          return;
-        }
-        auto to_hex = [](const std::vector<uint8_t>& buf) {
-          std::string str;
-          const char hex[] = "0123456789abcdef";
-          for (auto n : buf) {
-            str += hex[(n >> 3) & 0xf];
-            str += hex[n & 0xe];
-          }
-          return str;
-        };
-        std::stringstream ss;
-        ss << "SRTP_CLIENT_KEY " << to_hex(km->client_write_key) << "\n";
-        ss << "SRTP_CLIENT_SALT " << to_hex(km->client_write_salt) << "\n";
-        ss << "SRTP_SERVER_KEY " << to_hex(km->server_write_key) << "\n";
-        ss << "SRTP_SERVER_SALT " << to_hex(km->server_write_salt) << "\n";
-        std::ofstream ofs(config_.srtp_key_log_file, std::ios::app);
-        if (ofs) {
-          ofs.write(ss.str().c_str(), ss.str().size());
-        }
-        key_exported_ = true;
-      }
-    }
+    RTC_LOG(LS_INFO) << "OnNotify: " << text;
   }
   void OnPush(std::string text) override {}
   void OnMessage(std::string label, std::string data) override {}
@@ -294,7 +267,6 @@ class Sumomo : public std::enable_shared_from_this<Sumomo>,
   std::shared_ptr<sora::SoraSignaling> conn_;
   std::unique_ptr<boost::asio::io_context> ioc_;
   std::unique_ptr<SDLRenderer> renderer_;
-  bool key_exported_ = false;
 };
 
 void add_optional_bool(CLI::App& app,
@@ -349,7 +321,7 @@ int main(int argc, char* argv[]) {
 
   auto is_json = CLI::Validator(
       [](std::string input) -> std::string {
-        boost::json::error_code ec;
+        boost::system::error_code ec;
         boost::json::parse(input, ec);
         if (ec) {
           return "Value " + input + " is not JSON Value";
@@ -383,6 +355,7 @@ int main(int argc, char* argv[]) {
   app.add_option("--client-id", config.client_id, "Client ID");
   app.add_option("--video", config.video, "Send video to sora (default: true)");
   app.add_option("--audio", config.audio, "Send audio to sora (default: true)");
+  app.add_option("--video-device", config.video_device, "Video device name");
   app.add_option("--video-codec-type", config.video_codec_type,
                  "Video codec for send")
       ->check(CLI::IsMember({"", "VP8", "VP9", "AV1", "H264", "H265"}));
@@ -434,17 +407,36 @@ int main(int argc, char* argv[]) {
                "Use fullscreen window for videos");
   app.add_flag("--show-me", config.show_me, "Show self video");
 
-  // SRTP keying material の出力
-  app.add_option("--srtp-key-log-file", config.srtp_key_log_file,
-                 "SRTP keying material output file");
-
   // 証明書に関するオプション
   app.add_flag("--insecure", config.insecure, "Allow insecure connection");
-  app.add_option("--client-cert", config.client_cert, "Client certificate file")->check(CLI::ExistingFile);
-  app.add_option("--client-key", config.client_key, "Client key file")->check(CLI::ExistingFile);
-  app.add_option("--ca-cert", config.ca_cert, "CA certificate file")->check(CLI::ExistingFile);
+  app.add_option("--client-cert", config.client_cert, "Client certificate file")
+      ->check(CLI::ExistingFile);
+  app.add_option("--client-key", config.client_key, "Client key file")
+      ->check(CLI::ExistingFile);
+  app.add_option("--ca-cert", config.ca_cert, "CA certificate file")
+      ->check(CLI::ExistingFile);
+
+  // DegradationPreference に関するオプション
+  auto degradation_preference_map =
+      std::vector<std::pair<std::string, webrtc::DegradationPreference>>(
+          {{"disabled", webrtc::DegradationPreference::DISABLED},
+           {"maintain_framerate",
+            webrtc::DegradationPreference::MAINTAIN_FRAMERATE},
+           {"maintain_resolution",
+            webrtc::DegradationPreference::MAINTAIN_RESOLUTION},
+           {"balanced", webrtc::DegradationPreference::BALANCED}});
+  app.add_option("--degradation-preference", config.degradation_preference,
+                 "Degradation preference")
+      ->transform(CLI::CheckedTransformer(degradation_preference_map,
+                                          CLI::ignore_case));
 
   // SoraClientContextConfig に関するオプション
+  std::string audio_recording_device;
+  app.add_option("--audio-recording-device", audio_recording_device,
+                 "Recording device name");
+  std::string audio_playout_device;
+  app.add_option("--audio-playout-device", audio_playout_device,
+                 "Playout device name");
   std::optional<bool> use_hardware_encoder;
   add_optional_bool(app, "--use-hardware-encoder", use_hardware_encoder,
                     "Use hardware encoder");
@@ -477,6 +469,12 @@ int main(int argc, char* argv[]) {
   }
 
   auto context_config = sora::SoraClientContextConfig();
+  if (!audio_recording_device.empty()) {
+    context_config.audio_recording_device = audio_recording_device;
+  }
+  if (!audio_playout_device.empty()) {
+    context_config.audio_playout_device = audio_playout_device;
+  }
   if (use_hardware_encoder != std::nullopt) {
     context_config.use_hardware_encoder = *use_hardware_encoder;
   }