+++ include/sora/hwenc_vpl/vpl_video_decoder.h
@@ -2,6 +2,7 @@
 #define SORA_HWENC_VPL_VPL_VIDEO_DECODER_H_
 
 #include <memory>
+#include <optional>
 
 // WebRTC
 #include <api/video/video_codec_type.h>