+++ include/sora/websocket.h
@@ -6,6 +6,7 @@
 #include <optional>
 
 // Boost
+#include <boost/asio/deadline_timer.hpp>
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <boost/asio/ssl/context.hpp>