+++ src/hwenc_nvcodec/nvcodec_video_decoder.cpp
@@ -113,8 +113,8 @@ int32_t NvCodecVideoDecoder::Decode(const webrtc::EncodedImage& input_image,
                                            .set_video_frame_buffer(i420_buffer)
                                            .set_timestamp_rtp(pts)
                                            .build();
-    decode_complete_callback_->Decoded(decoded_image, absl::nullopt,
-                                       absl::nullopt);
+    decode_complete_callback_->Decoded(decoded_image, std::nullopt,
+                                       std::nullopt);
 
     decoder_->UnlockFrame(frame);
   }