+++ src/sora_signaling.cpp
@@ -15,7 +15,6 @@
 #include "sora/rtc_ssl_verifier.h"
 #include "sora/rtc_stats.h"
 #include "sora/session_description.h"
-#include "sora/srtp_keying_material_exporter.h"
 #include "sora/url_parts.h"
 #include "sora/version.h"
 #include "sora/zlib_helper.h"
@@ -619,26 +618,30 @@ void SoraSignaling::DoInternalDisconnect(
     }
   };
 
-  // Close 処理中に、意図しない場所で WS Close が呼ばれた場合の対策。
-  // 例えば dc_->Close()→送信完了して on_ws_close_ に値を設定して切断を待つ、
-  // となるまでの間に WS Close された場合、on_ws_close_ に値が設定されていなくて
-  // 永遠に終了できなくなってしまう。
-  if (ws_connected_) {
-    on_ws_close_ = [self = shared_from_this(),
-                    on_close](boost::system::error_code ec) {
-      boost::system::error_code tec;
-      self->closing_timeout_timer_.cancel(tec);
-      auto ws_reason = self->ws_->reason();
-      self->SendOnWsClose(ws_reason);
-      std::string message =
-          "Unintended disconnect WebSocket during Disconnect process: ec=" +
-          ec.message() + " wscode=" + std::to_string(ws_reason.code) +
-          " wsreason=" + ws_reason.reason.c_str();
-      on_close(false, SoraSignalingErrorCode::CLOSE_FAILED, message);
-    };
-  }
-
   if (using_datachannel_ && ws_connected_) {
+    // DC の切断タイムアウトがあるので、closing_timeout_timer_ は使わない
+    std::shared_ptr<bool> ws_close_called = std::make_shared<bool>(false);
+    std::shared_ptr<bool> dc_close_called = std::make_shared<bool>(false);
+    on_ws_close_ = [self = shared_from_this(), on_close, ws_close_called,
+                    dc_close_called](boost::system::error_code ec) {
+      // 既に DC 側で処理済み
+      if (*dc_close_called) {
+        return;
+      }
+      *ws_close_called = true;
+      auto reason = self->ws_->reason();
+      self->SendOnWsClose(reason);
+      if (ec != boost::beast::websocket::error::closed) {
+        std::string message = "Failed to close WebSocket: ec=" + ec.message() +
+                              " wscode=" + std::to_string(reason.code) +
+                              " wsreason=" + reason.reason.c_str();
+        on_close(false, SoraSignalingErrorCode::CLOSE_FAILED, message);
+      } else {
+        on_close(true, SoraSignalingErrorCode::CLOSE_SUCCEEDED,
+                 "Succeeded to close Websocket (DC signaling is enabled)");
+      }
+    };
+
     std::string text =
         force_error_code == std::nullopt
             ? R"({"type":"disconnect","reason":"NO-ERROR"})"
@@ -646,51 +649,32 @@ void SoraSignaling::DoInternalDisconnect(
     webrtc::DataBuffer disconnect = ConvertToDataBuffer("signaling", text);
     dc_->Close(
         disconnect,
-        [self = shared_from_this(), on_close, force_error_code,
-         message](boost::system::error_code ec1) {
-          self->closing_timeout_timer_.expires_from_now(
-              boost::posix_time::seconds(
-                  self->config_.websocket_close_timeout));
-          self->closing_timeout_timer_.async_wait(
-              [self](boost::system::error_code ec) {
-                if (ec) {
-                  return;
-                }
-                self->ws_->Cancel();
-              });
-          self->on_ws_close_ = [self, ec1,
-                                on_close](boost::system::error_code ec2) {
-            boost::system::error_code tec;
-            self->closing_timeout_timer_.cancel(tec);
-            auto ws_reason = self->ws_->reason();
-            self->SendOnWsClose(ws_reason);
-            std::string ws_reason_str =
-                " wscode=" + std::to_string(ws_reason.code) +
-                " wsreason=" + ws_reason.reason.c_str();
+        [self = shared_from_this(), on_close, ws_close_called,
+         dc_close_called](boost::system::error_code ec) {
+          // 正常な切断なら何もしない
+          if (!ec) {
+            return;
+          }
+          // 既に WS 側で処理済み
+          if (*ws_close_called) {
+            return;
+          }
+          *dc_close_called = true;
 
-            bool ec2_error = ec2 != boost::beast::websocket::error::closed;
-            bool succeeded = true;
-            std::string message =
-                "Succeeded to close DataChannel and Websocket";
-            auto error_code = SoraSignalingErrorCode::CLOSE_SUCCEEDED;
-            if (ec1 && ec2_error) {
-              succeeded = false;
-              message = "Failed to close DataChannel and WebSocket: ec1=" +
-                        ec1.message() + " ec2=" + ec2.message() + ws_reason_str;
-              error_code = SoraSignalingErrorCode::CLOSE_FAILED;
-            } else if (ec1 && !ec2_error) {
-              succeeded = false;
-              message = "Failed to close DataChannel (WS succeeded): ec=" +
-                        ec1.message() + ws_reason_str;
-              error_code = SoraSignalingErrorCode::CLOSE_FAILED;
-            } else if (!ec1 && ec2_error) {
-              succeeded = false;
-              message = "Failed to close WebSocket (DC succeeded): ec=" +
-                        ec2.message() + ws_reason_str;
-              error_code = SoraSignalingErrorCode::CLOSE_FAILED;
-            }
-            on_close(succeeded, error_code, message);
-          };
+          // DC 切断のタイムアウトかつ WebSocket の Close 処理が来てない状態なので何か問題が起きてる
+          self->ws_->Cancel();
+
+          // reason は自分で作る
+          auto reason = boost::beast::websocket::close_reason(
+              (boost::beast::websocket::close_code)4999,
+              "DISCONNECT-WAIT-TIMEOUT-ERROR");
+          self->SendOnWsClose(reason);
+
+          std::string message =
+              "Failed to close WebSocket (Close timeout): ec=" + ec.message() +
+              " wscode=" + std::to_string(reason.code) +
+              " wsreason=" + reason.reason.c_str();
+          on_close(false, SoraSignalingErrorCode::CLOSE_FAILED, message);
         },
         config_.disconnect_wait_timeout);
 
@@ -715,46 +699,56 @@ void SoraSignaling::DoInternalDisconnect(
     SendOnSignalingMessage(SoraSignalingType::DATACHANNEL,
                            SoraSignalingDirection::SENT, std::move(text));
   } else if (!using_datachannel_ && ws_connected_) {
+    closing_timeout_timer_.expires_from_now(
+        boost::posix_time::seconds(config_.websocket_close_timeout));
+    closing_timeout_timer_.async_wait(
+        [self = shared_from_this()](boost::system::error_code ec) {
+          if (ec) {
+            return;
+          }
+          self->ws_->Cancel();
+        });
+    on_ws_close_ = [self = shared_from_this(),
+                    on_close](boost::system::error_code ec) {
+      boost::system::error_code tec;
+      self->closing_timeout_timer_.cancel(tec);
+      auto reason = self->ws_->reason();
+      if (ec == boost::asio::error::operation_aborted) {
+        // タイムアウトによる切断なので 4999 をコールバックする
+        auto timeout_reason = boost::beast::websocket::close_reason(
+            (boost::beast::websocket::close_code)4999,
+            "DISCONNECT-WAIT-TIMEOUT-ERROR");
+        self->SendOnWsClose(timeout_reason);
+      } else {
+        self->SendOnWsClose(reason);
+      }
+      bool ec_error = ec != boost::beast::websocket::error::closed;
+      if (ec_error) {
+        on_close(false, SoraSignalingErrorCode::CLOSE_FAILED,
+                 "Failed to close WebSocket: ec=" + ec.message() +
+                     " wscode=" + std::to_string(reason.code) +
+                     " wsreason=" + reason.reason.c_str());
+        return;
+      }
+      on_close(true, SoraSignalingErrorCode::CLOSE_SUCCEEDED,
+               "Succeeded to close WebSocket (DC signaling is not enabled)");
+    };
     boost::json::value disconnect = {{"type", "disconnect"},
                                      {"reason", "NO-ERROR"}};
     WsWriteSignaling(
         boost::json::serialize(disconnect),
         [self = shared_from_this(), on_close](boost::system::error_code ec,
                                               std::size_t) {
-          if (ec) {
-            on_close(
-                false, SoraSignalingErrorCode::CLOSE_FAILED,
-                "Failed to write disconnect message to close WebSocket: ec=" +
-                    ec.message());
+          // 書き込みが成功した段階で Sora サーバーから切断されるので、
+          // 成功時の処理は不要
+          if (!ec) {
             return;
           }
 
-          self->closing_timeout_timer_.expires_from_now(
-              boost::posix_time::seconds(
-                  self->config_.websocket_close_timeout));
-          self->closing_timeout_timer_.async_wait(
-              [self](boost::system::error_code ec) {
-                if (ec) {
-                  return;
-                }
-                self->ws_->Cancel();
-              });
-          self->on_ws_close_ = [self, on_close](boost::system::error_code ec) {
-            boost::system::error_code tec;
-            self->closing_timeout_timer_.cancel(tec);
-            auto reason = self->ws_->reason();
-            self->SendOnWsClose(reason);
-            bool ec_error = ec != boost::beast::websocket::error::closed;
-            if (ec_error) {
-              on_close(false, SoraSignalingErrorCode::CLOSE_FAILED,
-                       "Failed to close WebSocket: ec=" + ec.message() +
-                           " wscode=" + std::to_string(reason.code) +
-                           " wsreason=" + reason.reason.c_str());
-              return;
-            }
-            on_close(true, SoraSignalingErrorCode::CLOSE_SUCCEEDED,
-                     "Succeeded to close WebSocket");
-          };
+          on_close(
+              false, SoraSignalingErrorCode::CLOSE_FAILED,
+              "Failed to write disconnect message to close WebSocket: ec=" +
+                  ec.message());
         });
   } else {
     on_close(false, SoraSignalingErrorCode::INTERNAL_ERROR,
@@ -847,60 +841,37 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
       RTC_LOG(LS_ERROR) << "OnRead: state is Closing but on_ws_close_ is null";
       return;
     }
-    if (state_ == State::Connected && !using_datachannel_) {
-      // 何かエラーが起きたので切断する
-      ws_connected_ = false;
-      auto error_code = ec == boost::beast::websocket::error::closed
-                            ? SoraSignalingErrorCode::WEBSOCKET_ONCLOSE
-                            : SoraSignalingErrorCode::WEBSOCKET_ONERROR;
-      auto reason = ws_->reason();
-      SendOnWsClose(reason);
-      std::string reason_str = " wscode=" + std::to_string(reason.code) +
-                               " wsreason=" + reason.reason.c_str();
-      SendOnDisconnect(error_code, "Failed to read WebSocket: ec=" +
-                                       ec.message() + reason_str);
-      return;
-    }
     if (state_ == State::Connected && using_datachannel_ && !ws_connected_) {
       // ignore_disconnect_websocket による WS 切断なので何もしない
       return;
     }
-    if (state_ == State::Connected && using_datachannel_ && ws_connected_) {
-      // DataChanel で reason: "WEBSOCKET-ONCLOSE" または "WEBSOCKET-ONERROR" を送る事を試みてから終了する
-      std::string text =
-          ec == boost::beast::websocket::error::closed
-              ? R"({"type":"disconnect","reason":"WEBSOCKET-ONCLOSE"})"
-              : R"({"type":"disconnect","reason":"WEBSOCKET-ONERROR"})";
-      webrtc::DataBuffer disconnect = ConvertToDataBuffer("signaling", text);
-      auto error_code = ec == boost::beast::websocket::error::closed
+    if (state_ == State::Connected) {
+      // 何かエラーが起きたので切断する
+      ws_connected_ = false;
+      auto reason = ws_->reason();
+      auto error_code = reason.code == 1000
+                            ? SoraSignalingErrorCode::CLOSE_SUCCEEDED
+                        : ec == boost::beast::websocket::error::closed
                             ? SoraSignalingErrorCode::WEBSOCKET_ONCLOSE
                             : SoraSignalingErrorCode::WEBSOCKET_ONERROR;
-      auto reason = ws_->reason();
       SendOnWsClose(reason);
       std::string reason_str = " wscode=" + std::to_string(reason.code) +
                                " wsreason=" + reason.reason.c_str();
-      state_ = State::Closing;
-      dc_->Close(
-          disconnect,
-          [self = shared_from_this(), ec, error_code,
-           reason_str](boost::system::error_code ec2) {
-            if (ec2) {
-              self->SendOnDisconnect(
-                  error_code,
-                  "Failed to read WebSocket and to close DataChannel: ec=" +
-                      ec.message() + " ec2=" + ec2.message() + reason_str);
-              return;
-            }
-            self->SendOnDisconnect(
-                error_code,
-                "Failed to read WebSocket and succeeded to close "
-                "DataChannel: ec=" +
-                    ec.message() + reason_str);
-          },
-          config_.disconnect_wait_timeout);
-
-      SendOnSignalingMessage(SoraSignalingType::DATACHANNEL,
-                             SoraSignalingDirection::SENT, std::move(text));
+      std::string message;
+      if (error_code == SoraSignalingErrorCode::CLOSE_SUCCEEDED) {
+        if (using_datachannel_) {
+          message =
+              "Succeeded to close WebSocket and DataChannel from server: ec=" +
+              ec.message() + reason_str;
+        } else {
+          message =
+              "Succeeded to close WebSocket from server: ec=" + ec.message() +
+              reason_str;
+        }
+      } else {
+        message = "Failed to read WebSocket: ec=" + ec.message() + reason_str;
+      }
+      SendOnDisconnect(error_code, message);
       return;
     }
 
@@ -1025,18 +996,21 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
               for (auto v : encodings_json) {
                 auto p = v.as_object();
                 webrtc::RtpEncodingParameters params;
-                // absl::optional<uint32_t> ssrc;
+                // std::optional<uint32_t> ssrc;
                 // double bitrate_priority = kDefaultBitratePriority;
-                // enum class Priority { kVeryLow, kLow, kMedium, kHigh };
                 // Priority network_priority = Priority::kLow;
-                // absl::optional<int> max_bitrate_bps;
-                // absl::optional<int> min_bitrate_bps;
-                // absl::optional<double> max_framerate;
-                // absl::optional<int> num_temporal_layers;
-                // absl::optional<double> scale_resolution_down_by;
+                // std::optional<int> max_bitrate_bps;
+                // std::optional<int> min_bitrate_bps;
+                // std::optional<double> max_framerate;
+                // std::optional<int> num_temporal_layers;
+                // std::optional<double> scale_resolution_down_by;
+                // std::optional<std::string> scalability_mode;
+                // std::optional<Resolution> scale_resolution_down_to;
                 // bool active = true;
                 // std::string rid;
+                // bool request_key_frame = false;
                 // bool adaptive_ptime = false;
+                // std::optional<RtpCodec> codec;
                 params.rid = p["rid"].as_string().c_str();
                 if (p.count("maxBitrate") != 0) {
                   params.max_bitrate_bps = p["maxBitrate"].to_number<int>();
@@ -1087,6 +1061,12 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
                     }
                   }
                 }
+                if (p.count("scaleResolutionDownTo") != 0) {
+                  auto& obj = p["scaleResolutionDownTo"].as_object();
+                  auto& v = params.scale_resolution_down_to.emplace();
+                  v.width = obj.at("maxWidth").to_number<int>();
+                  v.height = obj.at("maxHeight").to_number<int>();
+                }
                 encoding_parameters.push_back(params);
               }
 
@@ -1097,9 +1077,14 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
             SessionDescription::CreateAnswer(
                 self->pc_.get(),
                 [self](webrtc::SessionDescriptionInterface* desc) {
+                  if (self->config_.degradation_preference) {
+                    self->SetDegradationPreference(
+                        self->video_mid_,
+                        *self->config_.degradation_preference);
+                  }
+
                   std::string sdp;
                   desc->ToString(&sdp);
-                  //self->manager_->SetParameters();
                   boost::asio::post(*self->config_.io_context, [self, sdp]() {
                     if (!self->pc_) {
                       return;
@@ -1148,9 +1133,14 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
             SessionDescription::CreateAnswer(
                 self->pc_.get(),
                 [self, answer_type](webrtc::SessionDescriptionInterface* desc) {
+                  if (self->config_.degradation_preference) {
+                    self->SetDegradationPreference(
+                        self->video_mid_,
+                        *self->config_.degradation_preference);
+                  }
+
                   std::string sdp;
                   desc->ToString(&sdp);
-                  //self->manager_->SetParameters();
                   boost::asio::post(*self->config_.io_context,
                                     [self, sdp, answer_type]() {
                                       if (!self->pc_) {
@@ -1392,6 +1382,29 @@ void SoraSignaling::ResetEncodingParameters() {
     enc.ssrc = it->ssrc;
   }
   parameters.encodings = new_encodings;
+  sender->SetParameters(parameters);
+}
+
+void SoraSignaling::SetDegradationPreference(
+    std::string mid,
+    webrtc::DegradationPreference degradation_preference) {
+  rtc::scoped_refptr<webrtc::RtpTransceiverInterface> video_transceiver;
+  for (auto transceiver : pc_->GetTransceivers()) {
+    if (transceiver->mid() && *transceiver->mid() == mid) {
+      video_transceiver = transceiver;
+      break;
+    }
+  }
+
+  if (video_transceiver == nullptr) {
+    RTC_LOG(LS_ERROR) << "video transceiver not found";
+    return;
+  }
+
+  rtc::scoped_refptr<webrtc::RtpSenderInterface> sender =
+      video_transceiver->sender();
+  webrtc::RtpParameters parameters = sender->GetParameters();
+  parameters.degradation_preference = degradation_preference;
   sender->SetParameters(parameters);
 }