+++ CHANGES.md
@@ -13,25 +13,38 @@
 
 - [CHANGE] `boost::optional` を利用していた部分を全て `std::optional` に変更
   - @melpon
+- [CHANGE] SRTP keying material を取得する機能を削除
+  - @melpon
 - [UPDATE] VPL を v2.13.0 に上げる
   - @voluntas
 - [UPDATE] CMake を 3.30.5 にあげる
   - @voluntas
-- [UPDATE] Boost を 1.86.0 にあげる
-  - @voluntas
-- [UPDATE] libwebrtc を m129.6668.1.1 にあげる
+- [UPDATE] Boost を 1.87.0 にあげる
+  - boost::asio::deadline_timer が間接的にインクルードされなくなったので、websocket.h に boost/asio/deadline_timer.hpp のインクルードを追加して利用可能にする
+  - boost::json::error_code が削除されたので boost::system::error_code に置き換える
+  - @voluntas @torikizi @melpon
+- [UPDATE] libwebrtc を m132.6834.4.0 にあげる
   - H.265 Patch の修正に伴い、C++ SDK の H.265 に関する設定を変更
   - examples と test に `rtc::CreateRandomString` のヘッダを追加
   - `SetRtpTimestamp` を `frame.timestamp` から `frame.rtp_timestamp` に変更
+  - `scalable_track_source.h` と `scalable_track_source.cpp` の `absl::optional` を `std::optional` に変更
+  - `nvcodec_video_decoder.h` と `vpl_video_decoder.h` に `#include <optional>` を追加
+  - `nvcodec_video_decoder.cpp` と `vpl_video_decoder.cpp` の `absl::nullopt` を `std::nullopt` に変更
+  - Android の test アプリの cmake バージョンを VERSION と合わせる
+  - `requested_resolution` が `scale_resolution_down_to` に変更されたのでリネーム
+  - `ExportKeyingMaterial` が廃止されたため `ExportSrtpKeyingMaterial` に変更
+    - `ExportKeyingMaterial` は引数を 6 つ利用できたが `ExportSrtpKeyingMaterial` では 1 つしか利用できないため互換処理を追加
   - @tnoho @torikizi @melpon
 - [UPDATE] Xcode のバージョンを 15.4 にあげる
   - @tnoho
-- [UPDATE] SDL を 2.30.8 に上げる
+- [UPDATE] SDL を 2.30.11 に上げる
   - @torikizi
 - [UPDATE] BLEND2D_VERSION を上げる
   - @torikizi
 - [UPDATE] ASMJIT_VERSION を上げる
   - @torikizi
+- [UPDATE] OpenH264 を v2.5.0 に上げる
+  - @voluntas
 - [ADD] シグナリングメッセージを取得できるよう OnSignalingMessage を SoraSignalingObserver に追加する
   - @tnoho
 - [ADD] Intel VPL で AV1 エンコーダを動くようにする
@@ -39,8 +52,6 @@
 - [ADD] ルート証明書を指定可能にする
   - @melpon
 - [ADD] Ubuntu 24.04 armv8 に対応する
-  - @melpon
-- [ADD] VERSION と examples/VERSION のバージョンをチェックする仕組みを追加
   - @melpon
 - [ADD] WebSocket の Close を取得できるよう SendOnWsClose を SoraSignalingObserver に追加する
   - @tnoho
@@ -54,6 +65,18 @@
   - @melpon
 - [ADD] SoraSignalingConfig に forwarding_filters を追加
   - @melpon
+- [ADD] scaleResolutionDownTo に対応する
+  - @melpon
+- [ADD] SoraClientContext にオーディオデバイスの設定するオプションを追加
+  - @melpon
+- [ADD] sumomo にビデオデバイスとオーディオデバイスを設定するオプションを追加
+  - @melpon
+- [ADD] SoraSignalingConfig に degradation_preference を追加 
+  - @melpon
+- [ADD] sumomo と test/hello に degradation_preference を設定するオプションを追加
+  - @melpon
+- [ADD] WHIP に対応する
+  - @melpon
 - [FIX] HTTP Proxy 利用時の Websocket 初期化で insecure_ メンバ変数が初期化されていなかったのを修正
   - @melpon
 - [FIX] SoraSignalingConfig の client_cert と client_key に渡す必要がある値を、ファイルパスからファイルの内容に修正
@@ -61,9 +84,23 @@
   - @melpon
 - [FIX] SoraSignalingConfig の client_cert と client_key の型を `std::string` から `std::optional<std::string>` に修正
   - @melpon
+- [FIX] WS と DC が両方繋がっている時に切断した時、正常終了にも関わらずエラーが発生することがあるのを修正
+  - @melpon
+- [FIX] Sora から切断された場合に WS reason が 1000 だったら正常終了とする
+  - @melpon
+- [FIX] libdrm-dev と libva-dev の依存が不要になってたので削除
+  - @melpon
 
 ### misc
 
+- [CHANGE] GitHub Actions の ubuntu-latest を ubuntu-24.04 に変更
+  - @voluntas
+- [CHANGE] GitHub Actions の build に cron を設定する
+  - @voluntas
+- [UPDATE] Boost のダウンロード URL を変更する
+  - @voluntas
+- [UPDATE] CATCH2 を v3.7.1 にあげる
+  - @voluntas
 - [UPDATE] SDL2 を 2.30.7 にあげる
   - @voluntas
 - [ADD] sumomo に証明書に関するオプションを追加する
@@ -74,6 +111,8 @@
     - 指定できるファイルは PEM 形式
   - @melpon
 - [ADD] Ubuntu 24.04 armv8 向けの example を追加する
+  - @melpon
+- [ADD] VERSION と examples/VERSION のバージョンをチェックする仕組みを追加
   - @melpon
 
 ## 2024.7.0